requirements.txt
attrs
Automat
certifi
pandas
cffi
charset-normalizer
constantly
cryptography
cssselect
filelock
hyperlink
idna
incremental
itemadapter
itemloaders
jmespath
lxml
numpy
packaging
parsel
Protego
pyasn1
pyasn1-modules
pycparser
PyDispatcher
pyOpenSSL
python-dateutil
pytz
queuelib
requests
requests-file
service-identity
six
tldextract
typing-extensions
urllib3
w3lib
zope.interface
scrapy-user-agents
beautifulsoup4
Unidecode
scrapy-azure-exporter @ git+https://github.com/scrapy-plugins/scrapy-feedexporter-azure-storage@f7a296b554f3d4601dec115dc8d252f63212c4e8
pandas
scrapy-utils
webcolors
scrapeops_scrapy
scrapy_zyte_smartproxy
scrapy
Zyte-SmartProxy-Selenium
mysql-connector-python
mysql-handler
pandas
mypy-extensions
tqdm
pymysql
twisted
scrapy-inline-requests
azure-mgmt-compute
azure-identity
azure-common
azure-core
azure-storage-blob
scrapy-rotating-proxies
scrapeops-python-logger
scrapeops-python-requests
scrapeops-scrapy
scrapeops-scrapy-proxy-sdk
pillow
adal
inline
azure-storage-blob
scrapy-inline-requests
bs4
pydocumentdb
scrapoxy
scrapy-rotating-proxies
scrapeops_scrapy
scrapy-inline-requests
cloudscraper
urljoin
urlencoder
tldextract
aiohttp
asyncio
webcolors==1.11.1
nest_asyncio



